readded the production contingency for reseeding the database

+++ seeds/seed.js
@@ -5,7 +5,7 @@ const userData = require("./userData.json");
 const eventData = require("./eventData.json");
 
 const seedDatabase = async () => {
-  // if (process.env.NODE_ENV !== 'production') {
+  if (process.env.NODE_ENV !== 'production') {
     try {
       await sequelize.sync({ force: true }); // Only in development
 
@@ -24,9 +24,9 @@ const seedDatabase = async () => {
     } finally {
       process.exit(0);
     }
-  // } else {
-  //   console.log("Skipping seed in production.");
-  // }
+  } else {
+    console.log("Skipping seed in production.");
+  }
 };
 
 seedDatabase();
